+++ src/start.client.ts
@@ -4,9 +4,19 @@ import { automationClient } from "./automationClient";
 import { loadConfiguration } from "./configuration";
 import { enableDefaultScanning } from "./scan";
 
-loadConfiguration()
-    .then(configuration => {
-        enableDefaultScanning(configuration);
-        return configuration;
-    })
-    .then(configuration => automationClient(configuration).run());
+try {
+    loadConfiguration()
+        .then(configuration => {
+            enableDefaultScanning(configuration);
+            return configuration;
+        })
+        .then(configuration => automationClient(configuration).run())
+        .then(() => process.exit(0))
+        .catch(e => {
+            console.error(`Error: ${e.message}`);
+            process.exit(1);
+        });
+} catch (e) {
+    console.error(`Uncaught exception: ${e.message}`);
+    process.exit(10);
+}